Fixing the issue with wrong file name being passed to the GitLoader. Moving the GitConfig into a public class. Removing the @Component annotation from EtcdConfigurationRepositoryImpl so that the client can define the prefered repository implementation.
